requirements/requirements.txt
Django==5.1.4
django-cors-headers==4.2.0
django-filter==23.2
dj_database_url==2.1.0
django-simple-history==3.3.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
drf-yasg==1.21.7
python-decouple==3.8
gunicorn==20.1.0
watchdog==4.0.1
whitenoise==6.7.0
psycopg2==2.9.10
